python/dataset-experiments/requirements.txt
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
distro==1.9.0
exceptiongroup==1.3.0
galileo==1.14.0
galileo-core==3.59.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.2.6
openai==1.99.9
pandas==2.3.1
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvloop==0.21.0
wrapt==1.17.3
